requirements.txt
pandas
requests
websocket-client
flask